requirements.txt
apache-airflow-providers-amazon==2.2.0
apache-airflow-providers-google==5.0.0
apache-airflow-providers-postgres==1.0.2
apache-airflow-providers-snowflake==2.1.0
airflow-provider-great-expectations==0.0.8
dbt==0.21.0
